// packages/cosmic/src/accordion/interface.ts
/*
 * Copyright (c) Baidu, Inc. All Rights Reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface AccordionProps {
    /**
     * 是否允许同时存在多个展开的面板
     * @default true
     */
    multiple?: boolean;

    /**
     * 默认展开的折叠面板标识符。手风琴模式下为 string，非手风琴模式为 string[]
     * @default []
     */
    value?: string | string[];
};

export interface AccordionEvents {
    /**
     * 面板展开状态变化时触发
     */
    change: {
        value: string;
        expanded: boolean;
        event: Event;
    };
}

export type AccordionData = Required<AccordionProps>;